actions/convolution.py
from PIL import Image
from io import BytesIO
import numpy

# g. Convolução entre uma imagem f e uma máscara h, com offset, tal que 
# g = f*h+offset. O sistema deverá permitir que o usuário especifique a 
# imagem f e, via interface gráfica ou arquivo com formatação específica, 
# os valores dos elementos de h (n x m). Considere nulos os valores 
# indefinidos de fi,j e hi,j.
def convolution(image):
  PILimage = Image.open(BytesIO(image))
  if (PILimage.mode == "RGB"):
    (R, G, B) = PILimage.split()
    R = numpy.asarray(R)
    G = numpy.asarray(G)
    B = numpy.asarray(B)

    # FAÇA AS ALTERAÇÕES EM R, G E B

    # END FAÇA AS ALTERAÇÕES EM R, G E B

    NPimage = numpy.zeros([PILimage.height, PILimage.width, 3], dtype=numpy.uint8)
    NPimage[:, :, 0] = R
    NPimage[:, :, 1] = G
    NPimage[:, :, 2] = B
  else:
    NPimage = numpy.asarray(PILimage.convert("L"))

    # FAÇA AS ALTERAÇÕES EM NPimage

    # END FAÇA AS ALTERAÇÕES EM NPimage

  # Transforma o array numpy (que é uma matriz) novamente em uma imagem
  PILimage = Image.fromarray(numpy.uint8(NPimage))
  return PILimage